# docs/swagger.yaml
definitions:
...
  http.TripData:
    properties:
      city_id:
        type: integer
      description:
        type: string
      end_date:
        type: string
      name:
        type: string
      private:
        type: boolean
      start_date:
        type: string
      user_id:
        type: integer
    type: object
  httpresponses.ErrorResponse:
    properties:
      message:
        type: string
    type: object
...
  models.Trip:
    properties:
      city_id:
        type: integer
      created_at:
        type: string
      description:
        type: string
      end_date:
        type: string
      id:
        type: integer
      name:
        type: string
      private:
        type: boolean
      start_date:
        type: string
      user_id:
        type: integer
    type: object
...
info:
  contact: {}
paths:
...
  /trips:
    post:
      consumes:
      - application/json
      description: Create a new trip with given fields
      parameters:
      - description: Trip details
        in: body
        name: tripData
        required: true
        schema:
          $ref: '#/definitions/http.TripData'
      produces:
      - application/json
      responses:
        "201":
          description: Trip created successfully
          schema:
            $ref: '#/definitions/models.Trip'
        "400":
          description: Invalid request
          schema:
            $ref: '#/definitions/httpresponses.ErrorResponse'
        "404":
          description: Invalid request
          schema:
            $ref: '#/definitions/httpresponses.ErrorResponse'
        "500":
          description: Failed to create trip
          schema:
            $ref: '#/definitions/httpresponses.ErrorResponse'
      summary: Create a new trip
  /trips/{id}:
    delete:
      description: Delete a trip by trip ID
      parameters:
      - description: Trip ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "204":
          description: Trip deleted successfully
        "400":
          description: Invalid trip ID
          schema:
            $ref: '#/definitions/httpresponses.ErrorResponse'
        "404":
          description: Trip not found
          schema:
            $ref: '#/definitions/httpresponses.ErrorResponse'
        "500":
          description: Failed to delete trip
          schema:
            $ref: '#/definitions/httpresponses.ErrorResponse'
      summary: Delete a trip
    get:
      description: Get trip details by trip ID
      parameters:
      - description: Trip ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Trip details
          schema:
            $ref: '#/definitions/models.Trip'
        "400":
          description: Invalid trip ID
          schema:
            $ref: '#/definitions/httpresponses.ErrorResponse'
        "404":
          description: Trip not found
          schema:
            $ref: '#/definitions/httpresponses.ErrorResponse'
        "500":
          description: Failed to retrieve trip
          schema:
            $ref: '#/definitions/httpresponses.ErrorResponse'
      summary: Retrieve a trip by ID
    post:
      description: Add a place with given place_id to a trip
      parameters:
      - description: Trip ID
        in: path
        name: id
        required: true
        type: integer
      - description: Place ID
        in: body
        name: place_id
        required: true
        schema:
          type: integer
      produces:
      - application/json
      responses:
        "201":
          description: Place added to trip successfully
        "400":
          description: Invalid place ID
          schema:
            $ref: '#/definitions/httpresponses.ErrorResponse'
        "404":
          description: Place not found
          schema:
            $ref: '#/definitions/httpresponses.ErrorResponse'
        "500":
          description: Failed to add place to trip
          schema:
            $ref: '#/definitions/httpresponses.ErrorResponse'
      summary: Add a place to a trip
    put:
      consumes:
      - application/json
      description: Update trip details by trip ID
      parameters:
      - description: Trip ID
        in: path
        name: id
        required: true
        type: integer
      - description: Updated trip details
        in: body
        name: tripData
        required: true
        schema:
          $ref: '#/definitions/http.TripData'
      produces:
      - application/json
      responses:
        "200":
          description: Trip updated successfully
          schema:
            $ref: '#/definitions/models.Trip'
        "400":
          description: Invalid trip data
          schema:
            $ref: '#/definitions/httpresponses.ErrorResponse'
        "404":
          description: Trip not found
          schema:
            $ref: '#/definitions/httpresponses.ErrorResponse'
        "500":
          description: Failed to update trip
          schema:
            $ref: '#/definitions/httpresponses.ErrorResponse'
      summary: Update an existing trip
...
  /users/{userID}/trips:
    get:
      description: Get all trips for a specific user
      parameters:
      - description: User ID
        in: path
        name: userID
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: List of trips
          schema:
            items:
              $ref: '#/definitions/models.Trip'
            type: array
        "400":
          description: Invalid user ID
          schema:
            $ref: '#/definitions/httpresponses.ErrorResponse'
        "404":
          description: Trips not found
          schema:
            $ref: '#/definitions/httpresponses.ErrorResponse'
        "500":
          description: Failed to retrieve trips
          schema:
            $ref: '#/definitions/httpresponses.ErrorResponse'
      summary: Retrieve trips by user ID
...
swagger: "2.0"